+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Unreleased
 - Added `distance_threshold` to `pairwise_link_collision_info` and `pairwise_link_collision` to allow collision checking given a penetration threshold
 - Added `sweep_collision_fn` to `birrt` to allow sweep collision check in edge expansion
 - Added `coarse_waypoints` to the `smooth_path` function to give options for use refined shortcut to ensure collision-free results.
+- Added `get_body_collision_vertices` for getting body collision vertices in its current configuration.
 
 **Changed**
 - Apply `HideOutput` to pybullet IK error printouts in `inverse_kinematics_helper`